Allow `set_storage` for `AllowMultiAssetPools` / `LiquidityWithdrawalFee`
