+++ parachains/runtimes/assets/test-utils/src/test_cases.rs
@@ -33,6 +33,9 @@ use sp_runtime::{
 use xcm::latest::prelude::*;
 use xcm_executor::{traits::ConvertLocation, XcmExecutor};
 
+// Re-export test_case from `parachains-runtimes-test-utils`
+pub use parachains_runtimes_test_utils::test_cases::change_storage_constant_by_governance_works;
+
 /// Test-case makes sure that `Runtime` can receive native asset from relay chain
 /// and can teleport it back and to the other parachains
 pub fn teleports_for_native_asset_works<
